move conftest.py inside test directory
